backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
numpy
scikit-learn
pydantic
python-multipart
